arreglando endpoints y fetch en proyectos

+++ Backend/apis/src/app/api/proyectos/[id]/route.ts
@@ -1,10 +1,28 @@
 import { NextResponse } from "next/server";
 import { DeleteProyectosController, PutProyectosController } from "@/controllers/proyectos/proyectosController";
+import { addLogEntry } from "@/models/admin/logModel";
+import { cookies } from "next/headers";
+import { verifyAccessToken, getUserIdFromSessionToken } from "@/lib/auth/login_tools";
 
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
   try {
+    const jar = await cookies();  
+    const token = jar.get("access_token")?.value;
+    if (!token) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    if (!(await verifyAccessToken(token))) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    const userId = await getUserIdFromSessionToken(token || "");
+    if (!userId) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    
     const { id } = params;
     const response = await DeleteProyectosController(id);
+    await addLogEntry(userId, "deleteProyecto", `Proyecto eliminado con ID: ${params.id}`);
     return NextResponse.json(
       { message: "Noticia eliminada correctamente", response },
       { status: 200 }
@@ -20,9 +38,23 @@ export async function DELETE(request: Request, { params }: { params: { id: strin
 
 export async function PUT(request: Request, { params }: { params: { id: string } }) {
   try {
+    const jar = await cookies();  
+    const token = jar.get("access_token")?.value;
+    if (!token) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    if (!(await verifyAccessToken(token))) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    const userId = await getUserIdFromSessionToken(token || "");
+    if (!userId) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    
     const { id } = params;
     const body = await request.formData();
-
+    await addLogEntry(userId, "updateProyecto", `Proyecto actualizado con ID: ${params.id}`);
     const response = await PutProyectosController(id, body);
     return NextResponse.json(
       { message: "Noticia actualizada correctamente", response },
